requirements.txt
selenium==4.11.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0 